# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8000
info:
  contact: {}
  description: Backend REST API for Jeffries Homeapp.
  title: Jeffries Homeapp API
  version: "1.0"
paths:
...
  /settings/backup:
    get:
      description: Generates a JSON backup of all user data
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: userId required
          schema:
            type: string
      summary: Export user data
      tags:
      - Settings
...
  /settings/telegram/status:
    get:
      description: Returns connection status and configuration info for the Telegram
        bot
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Telegram client not configured
          schema:
            type: string
      summary: Get Telegram bot status
      tags:
      - Settings
  /sync/calendar:
    post:
      description: Triggers a manual sync of Google Calendar to fetch schedules and
        personal events
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok, scheduleCount, personalCount, message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: userId required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Sync Calendar
      tags:
      - Sync
  /sync/gmail:
    post:
      description: Triggers a manual sync of Gmail messages
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: userId required
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Sync Gmail
      tags:
      - Sync
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"